requirements.txt
crispy-bootstrap5==2023.10
dj-database-url==2.1.0
django-crispy-forms==2.1
Django==4.2.7
gunicorn==21.2.0
Pillow==11.1.0
psycopg2-binary==2.9.7
python-dotenv==1.0.1
whitenoise==6.6.0